+++ database/migrations/2019_06_03_173501_create_prices_table.php
@@ -15,13 +15,9 @@ class CreatePricesTable extends Migration
     {
         Schema::create('prices', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('horeca')->nullable();
-            $table->integer('purchase');
-            $table->integer('distribution')->nullable();
-
-            $table->integer('horeca_unit')->nullable();
-            $table->integer('purchase_unit')->nullable();
-            $table->integer('distribution_unit')->nullable();
+            $table->unsignedDecimal('horeca')->nullable();
+            $table->unsignedDecimal('purchase');
+            $table->unsignedDecimal('distribution')->nullable();
 
             $table->unsignedDecimal('discount')->nullable();
             $table->unsignedDecimal('margin')->nullable();