+++ resources/views/beer/index.blade.php
@@ -47,8 +47,6 @@
         </div>
 
         @foreach($beers as $beer)
-
-
             <div class="row align-items-center mb-0 mt-0">
                 <div class="col-sm mb-0 mt-0">
                     <h5 class="text-primary" data-toggle="collapse" href={{ "#beer".$beer->id }} aria-expanded="false" aria-controls={{ "beer".$beer->id }}>{{ $beer->name }} <small class="text-secondary"> - {{ $beer->brewery->name }}</small></h5>
@@ -63,24 +61,25 @@
                 </div>
 
 
+
                 @hasanyrole('Publican|Admin')
                     <div class="col-sm-auto">
                         <h6 class="text-body">&euro; {{ $beer->price  ? $beer->price->distribution : 'n/d'}} {{ ($beer->price && $beer->packaging->type=='fusti' ) ? '- €/lt '.$beer->price->distributionLiter : ' '}}
-                            {{ ($beer->price && $beer->packaging->type=='bottiglie' ) ? '- €/bt '.$beer->price->distribution_unit : ' '}} (+Iva)</h6>
+                            {{ ($beer->price && $beer->packaging->type=='bottiglie' ) ? '- €/bt '.$beer->price->distribution_unit : ' '}}</h6>
                     </div>
                 @endhasanyrole
 
                 <div class="col-sm-auto">
+                <a class="text-primary" data-toggle="collapse" href={{ "#beer".$beer->id }}  aria-expanded="false" aria-controls={{ "beer".$beer->id }} >Espandi</a>
                 @hasrole('Admin')
                         <a href="/beers/{{ $beer->id }}/duplicate?packaging={{ request()->packaging }} " class="btn-primary">Duplica</a>
                         <a href="/beers/{{ $beer->id }}/edit?packaging={{ request()->packaging }} " class="btn-primary">Modifica</a>
                         <a href="/beers/{{ $beer->id }}/delete?packaging={{ request()->packaging }} " class="btn-danger">Elimina</a>
                 @endhasrole
-                    <a class="text-primary" data-toggle="collapse" href={{ "#beer".$beer->id }}  aria-expanded="false" aria-controls={{ "beer".$beer->id }} >Espandi</a>
                 </div>
 
-                <div class="card collapse" id={{ "beer".$beer->id }}>
-                    <div class="card-body">
+                <div class="collapse ml-2 mb-0 mt-0 mr-2 align-items-center" id={{ "beer".$beer->id }}>
+                    <div class="card card-body p-1">
                         {{ $beer->description }}
                     </div>
                 </div>